add a simple test suite
