+++ telml-bin/telml-bin.cabal
@@ -13,11 +13,8 @@ build-type: Simple
 cabal-version: 1.14
 tested-with:            GHC == 8.8.4, GHC == 8.10.7, GHC == 9.0.2, GHC == 9.2.4, GHC == 9.4.2
 
-executable telml
-  hs-source-dirs: src
-  main-is: Main.hs
-  default-language: Haskell2010
-  default-extensions: ScopedTypeVariables
+library
+  exposed-modules: Data.TeLML.Driver
   ghc-options: -Wall
   build-depends: base >=4.7 && <5
                , bytestring
@@ -25,3 +22,23 @@ executable telml
                , telml-markup
                , text
                , hslua-core
+  default-language: Haskell2010
+
+executable telml
+  hs-source-dirs: src
+  main-is: Main.hs
+  default-language: Haskell2010
+  default-extensions: ScopedTypeVariables
+  ghc-options: -Wall
+  build-depends: base >=4.7 && <5
+               , telml-bin
+
+test-suite telml-bin-tests
+  type: exitcode-stdio-1.0
+  main-is: TelmlBinTests.hs
+  build-depends: base
+               , hspec
+               , telml-bin
+               , bytestring
+               , text
+  hs-source-dirs: test
